Struct compiler::components::lexer::lexer::Lexed

pub struct Lexed;

Lexed is the result of tokenizing a parsable

Lexed | Lexeme Heap

More specifically Lexed is the structure that holds a structured form of the modules and lines of eureka source code. To break it down further its in a tree-like format. From top to bottom we will call it the y axis and from left to right the x axis. This is the way Lexed or (Lexeme Heap) should be traversed respectively:

Main Module (Root, the special module called main that is always on top of the Lexeme Heap)

Relatives of Modules (A collection of modules in the same y), /// - Line Relatives (A collection of lines in the same y)

Line families (A collection of lines in the same y as the parent of the specific y(x) line)

Implementations

impl Lexed

pub fn lex<T: AsRef<str>>(input: T) -> Self