Module compiler::scanner::scanner

Structs

CHAR_LITERAL_REGEX 🔒

Matches character literals.

IDENT_REGEX 🔒

Matches identifiers that start with underscores or ascii letters.

IDENT_STATE_MACHINE 🔒

State machine for validating identifiers.

NUMBER_LITERAL_REGEX 🔒

Matches zero and non-zero signless numbers.

NUMBER_STATE_MACHINE 🔒

State machine for validating numeric literals

Program

Source file split into its tokens, identifiers and constants.

STR_LITERAL_REGEX 🔒

Matches string literals.

Functions

parse_i32

Parses the input into a 32-bit signed integer.