Module compiler::state_machine

State machine implementation.

Structs

SerializedStateMachine ğŸ”’

SerializedTrans ğŸ”’

StateMachine

Implements the logic required for a finite state machine.

Enums

StateMachineDeserError

Errors the may be returned when deserializing a state machine.

Type Definitions

State ğŸ”’

StateArc ğŸ”’

Symbol ğŸ”’

SymbolArc ğŸ”’